requirements.txt
httpx==0.27.2
streamlit==1.37.1
python-dotenv==1.0.1
requests==2.32.3
groq==0.9.0
duckduckgo-search==6.2.12
wikipedia==1.4.0
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.0.1